fix: align rich output with current main
