+++ apps/web/src/locales/pt-pt/settings.json
@@ -28,6 +28,8 @@
   "agentGeneratedTaskTitles": "Títulos de tarefa gerados pelo agente",
   "agentGeneratedTaskTitlesDescription": "Quando está ativado, os formulários de nova tarefa e de subtarefa ocultam o campo manual de título e exigem um prompt. O Kandev mostra de imediato um título provisório com as primeiras seis palavras do prompt e depois pede ao agente que o substitua por um título conciso, com cerca de três palavras. Se o agente não conseguir mudar o nome da tarefa, o título provisório continua disponível para edição. As tarefas existentes e editadas mantêm o seu campo de título normal.",
   "agentGeneratedTaskTitlesDisabledHint": "Desativado por pré-definição; guarde esta definição antes de abrir uma janela de criação.",
+  "animateRichOutputCharts": "Animar gráficos apresentados pelos agentes",
+  "animateRichOutputChartsDescription": "Anime os gráficos de linhas e de barras quando os agentes os apresentam. Desative esta opção para reduzir o trabalho do navegador; os dados e os controlos dos gráficos continuam disponíveis. A definição de redução de movimento do sistema operativo tem sempre prioridade. Guardado neste dispositivo.",
   "apiTokenIsConfigured": "O token de API está configurado.",
   "appearance": "Aspeto",
   "appriseNotInstalled": "O Apprise ainda não está instalado. Pode adicioná-lo mais tarde para ativar as notificações remotas. <2>Ver as instruções de instalação</2>.",
@@ -189,6 +191,7 @@
   "discoveryAliasesNotificationSound": "áudio de notificação,sino,som de alerta",
   "discoveryAliasesPreferredShell": "shell,zsh,bash,shell do terminal",
   "discoveryAliasesResourceMetrics": "métricas,CPU,memória,disco,barra de estado",
+  "discoveryAliasesRichOutputMotion": "animação de gráficos,grafos,desempenho,movimento reduzido",
   "discoveryAliasesStartupPage": "página inicial,página de arranque,página de lançamento",
   "discoveryAliasesSubmitShortcut": "tecla de envio,enter para enviar,command enter",
   "discoveryAliasesTerminalFont": "família de letra,tipo de letra,monoespaçado",
@@ -503,6 +506,7 @@
   "resetToDefault": "Repor a pré-definição",
   "resizeRepositionOrMergeTheSelected": "Redimensionar, reposicionar ou fundir a divisão selecionada.",
   "resourceMetrics": "Métricas de recursos",
+  "richOutputChartMotion": "Movimento dos gráficos apresentados pelo agente",
   "restoreTheOriginalBuiltInLayout": "Restaurar o esquema incorporado original e descartar a sua substituição.",
   "resumeTheMostRecentlyOpenedTask": "Retomar a tarefa aberta mais recentemente nesta área de trabalho, neste dispositivo.",
   "retrySave": "Repetir a gravação",
