+++ apps/web/src/locales/zh-cn/settings.json
@@ -29,6 +29,8 @@
   "agentGeneratedTaskTitles": "由智能体生成任务标题",
   "agentGeneratedTaskTitlesDescription": "启用后，新建任务和子任务表单会隐藏手动标题字段，并要求填写提示词。Kandev 会立即使用提示词的前六个词生成临时标题，然后让智能体将其替换为约三个词的简洁标题。如果智能体无法重命名任务，你仍可编辑临时标题。现有任务和经过编辑的任务会继续显示常规标题字段。",
   "agentGeneratedTaskTitlesDisabledHint": "默认关闭；请先保存此设置，再打开创建对话框。",
+  "animateRichOutputCharts": "为智能体展示的图表启用动画",
+  "animateRichOutputChartsDescription": "智能体展示折线图和条形图时播放动画。关闭此选项可减少浏览器负载；图表数据和控件仍然可用。操作系统的“减少动态效果”设置始终优先。此设置保存在当前设备上。",
   "apiTokenIsConfigured": "API 令牌已配置。",
   "appearance": "外观",
   "appriseNotInstalled": "Apprise 尚未安装。你可以稍后添加它，以启用远程通知。<2>查看安装说明</2>。",
@@ -192,6 +194,7 @@
   "discoveryAliasesNotificationSound": "通知音,提示音,警报音",
   "discoveryAliasesPreferredShell": "shell,zsh,bash,终端 shell",
   "discoveryAliasesResourceMetrics": "指标,CPU,内存,磁盘,状态栏",
+  "discoveryAliasesRichOutputMotion": "图表动画,图形,性能,减少动态效果",
   "discoveryAliasesSettingsMenu": "侧边栏,导航,树形菜单,单分支展开,保留展开,菜单",
   "discoveryAliasesStartupPage": "主页,起始页,启动页",
   "discoveryAliasesStatusBar": "底栏,状态面板,连接状态,外观",
@@ -516,6 +519,7 @@
   "resetToDefault": "恢复默认设置",
   "resizeRepositionOrMergeTheSelected": "调整所选拆分区域的大小、位置或合并它。",
   "resourceMetrics": "资源指标",
+  "richOutputChartMotion": "智能体图表动态效果",
   "restoreTheOriginalBuiltInLayout": "恢复原始内置布局并舍弃其覆盖设置。",
   "resumeTheMostRecentlyOpenedTask": "在此设备上恢复此工作区中最近打开的任务。",
   "retrySave": "重试保存",
