+++ apps/web/src/locales/zh-hk/task.json
@@ -1147,10 +1147,10 @@
   "richOutputLoadingPreview": "正在載入預覽",
   "richOutputOpenFile": "開啟檔案",
   "richOutputPreview": "預覽",
-  "richOutputPreviewFailed": "預覽無法使用。",
+  "richOutputPreviewFailed": "預覽不可用。",
   "richOutputPreviewTruncated": "預覽已截斷。",
-  "richOutputTool": "Kandev：視覺化輸出",
-  "richOutputUnavailable": "此展示內容無法使用。",
+  "richOutputTool": "Kandev：可視化輸出",
+  "richOutputUnavailable": "此演示內容不可用。",
   "review": "審閱",
   "reviewComments": "審閱評論",
   "reviewCommentsReadyToSend_one": "{{count}} 條審閱評論待傳送",
