+++ apps/web/src/locales/zh-tw/settings.json
@@ -29,6 +29,8 @@
   "agentGeneratedTaskTitles": "由代理程式產生任務標題",
   "agentGeneratedTaskTitlesDescription": "啟用後，新建任務和子任務表單會隱藏手動標題欄位，並要求填寫提示詞。Kandev 會立即使用提示詞的前六個詞產生臨時標題，然後讓代理程式將其替換為約三個詞的簡潔標題。如果代理程式無法重新命名任務，你仍可編輯臨時標題。現有任務和經過編輯的任務會繼續顯示常規標題欄位。",
   "agentGeneratedTaskTitlesDisabledHint": "預設關閉；請先儲存此設定，再開啟建立對話方塊。",
+  "animateRichOutputCharts": "為代理程式展示的圖表啟用動畫",
+  "animateRichOutputChartsDescription": "代理程式展示折線圖和條形圖時播放動畫。關閉此選項可減少瀏覽器負載；圖表資料和控制項仍然可用。作業系統的「減少動態效果」設定始終優先。此設定儲存在目前裝置上。",
   "apiTokenIsConfigured": "API 權杖已設定。",
   "appearance": "外觀",
   "appriseNotInstalled": "Apprise 尚未安裝。你可以稍後新增它，以啟用遠端通知。<2>檢視安裝說明</2>。",
@@ -192,6 +194,7 @@
   "discoveryAliasesNotificationSound": "通知音,提示音,警報音",
   "discoveryAliasesPreferredShell": "shell,zsh,bash,終端 shell",
   "discoveryAliasesResourceMetrics": "指標,CPU,記憶體,磁碟,狀態列",
+  "discoveryAliasesRichOutputMotion": "圖表動畫,圖形,效能,減少動態效果",
   "discoveryAliasesSettingsMenu": "側邊欄,導覽,樹形選單,單分支展開,保留展開,選單",
   "discoveryAliasesStartupPage": "主頁,起始頁,啟動頁",
   "discoveryAliasesStatusBar": "底欄,狀態面板,連線狀態,外觀",
@@ -516,6 +519,7 @@
   "resetToDefault": "恢復預設設定",
   "resizeRepositionOrMergeTheSelected": "調整所選拆分割槽域的大小、位置或合併它。",
   "resourceMetrics": "資源指標",
+  "richOutputChartMotion": "代理程式圖表動態效果",
   "restoreTheOriginalBuiltInLayout": "恢復原始內建佈局並捨棄其覆蓋設定。",
   "resumeTheMostRecentlyOpenedTask": "在此裝置上恢復此工作區中最近開啟的任務。",
   "retrySave": "重試儲存",
