+++ apps/web/src/locales/zh-tw/task.json
@@ -1142,15 +1142,15 @@
   "retrying": "正在重試…",
   "reverseISearch": "(reverse-i-search)",
   "revertCommit": "還原提交",
-  "richOutputBinaryPreviewUnavailable": "無法預覽此檔案類型。",
+  "richOutputBinaryPreviewUnavailable": "無法預覽此檔案型別。",
   "richOutputHidePreview": "隱藏預覽",
   "richOutputLoadingPreview": "正在載入預覽",
   "richOutputOpenFile": "開啟檔案",
   "richOutputPreview": "預覽",
-  "richOutputPreviewFailed": "預覽無法使用。",
+  "richOutputPreviewFailed": "預覽不可用。",
   "richOutputPreviewTruncated": "預覽已截斷。",
   "richOutputTool": "Kandev：視覺化輸出",
-  "richOutputUnavailable": "此展示內容無法使用。",
+  "richOutputUnavailable": "此演示內容不可用。",
   "review": "審閱",
   "reviewComments": "審閱評論",
   "reviewCommentsReadyToSend_one": "{{count}} 條審閱評論待傳送",
